flask_factory_pattern/requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.3
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
SQLAlchemy==2.0.29
tomli==2.0.1
typing_extensions==4.11.0
Werkzeug==3.0.2
zipp==3.18.1
